test(output): move temp-file failure test behind !wasm tag
createTempFile is defined only in the native build (output.go is
//go:build !wasm), but the test referencing it lived in the untagged
output_test.go, breaking the wasm test build. Move it to a !wasm-tagged
file alongside the symbol it exercises.
